[SYCL] Separate OpenCL specific constructs from lit tests.
Adding 'opencl' requirement disabled several tests some of which
have non-opencl constructs as well. This patch separates those tests
to enable testing of SYCL core functionalities.
