+++ sycl/test/inorder_queue/in_order_buffs.cpp
@@ -1,5 +1,3 @@
-// REQUIRES: opencl
-
 // RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple %s -o %t.out -L %opencl_libs_dir -lOpenCL
 // RUN: %ACC_RUN_PLACEHOLDER %t.out
 // RUN: %CPU_RUN_PLACEHOLDER %t.out
@@ -17,14 +15,6 @@
 using namespace cl::sycl;
 
 const int dataSize = 32;
-
-bool isQueueInOrder(cl_command_queue cq) {
-  cl_command_queue_properties reportedProps;
-  cl_int iRet = clGetCommandQueueInfo(
-      cq, CL_QUEUE_PROPERTIES, sizeof(reportedProps), &reportedProps, nullptr);
-  assert(CL_SUCCESS == iRet && "Failed to obtain queue info from ocl device");
-  return (!(reportedProps & CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE));
-}
 
 int main() {
   int dataA[dataSize] = {0};
@@ -62,18 +52,6 @@ int main() {
 
       cgh.parallel_for<class ordered_reader>(myRange, myKernel);
     });
-
-    bool result = true;
-    cl_command_queue cq = Queue.get();
-    device dev = Queue.get_device();
-    bool expected_result = dev.is_host() ? true : isQueueInOrder(cq);
-
-    if (expected_result != result) {
-      std::cout << "Resulting queue order is OOO but expected order is inorder"
-                << std::endl;
-
-      return -1;
-    }
 
     auto readBufferB = bufB.get_access<access::mode::read>();
     for (size_t i = 0; i != dataSize; ++i) {
